+++ authapp/templates/authapp/notifications.html
@@ -12,7 +12,7 @@
                 {% for item in notifications_not_read %}
                     <div class="block-notification">
                             {{ item.add_datetime }}
-                            <a href="{% url 'auth:profile' %}">
+                            <a href="{% url 'profile:main' %}">
                                 <span class="font-weight-bold">
                                     {{item.sender }}
                                 </span>
@@ -36,7 +36,7 @@
                 {% for item in notifications_read %}
                     <div class="block-notification">
                             {{ item.add_datetime }}
-                            <a href="{% url 'auth:profile' %}">
+                            <a href="{% url 'profile:main' %}">
                                 <span class="font-weight-bold">
                                     {{item.sender }}
                                 </span>
