+++ db.json
@@ -54,43 +54,43 @@
  }
 },
 {
- "model": "mainapp.buttonsinprofile",
+ "model": "userprofile.buttonsinprofile",
  "pk": 1,
  "fields": {
   "name": "edit",
-  "include_html_file_name": "authapp/includes/inc__btn_edit.html"
+  "include_html_file_name": "userprofile/includes/inc__btn_edit.html"
  }
 },
 {
- "model": "mainapp.buttonsinprofile",
+ "model": "userprofile.buttonsinprofile",
  "pk": 2,
  "fields": {
   "name": "moderation",
-  "include_html_file_name": "authapp/includes/inc__btn_moderation.html"
+  "include_html_file_name": "userprofile/includes/inc__btn_moderation.html"
  }
 },
 {
- "model": "mainapp.buttonsinprofile",
+ "model": "userprofile.buttonsinprofile",
  "pk": 3,
  "fields": {
   "name": "delete",
-  "include_html_file_name": "authapp/includes/inc__btn_delete.html"
+  "include_html_file_name": "userprofile/includes/inc__btn_delete.html"
  }
 },
 {
- "model": "mainapp.buttonsinprofile",
+ "model": "userprofile.buttonsinprofile",
  "pk": 4,
  "fields": {
   "name": "draft",
-  "include_html_file_name": "authapp/includes/inc__btn_draft.html"
+  "include_html_file_name": "userprofile/includes/inc__btn_draft.html"
  }
 },
 {
- "model": "mainapp.buttonsinprofile",
+ "model": "userprofile.buttonsinprofile",
  "pk": 5,
  "fields": {
   "name": "recover",
-  "include_html_file_name": "authapp/includes/inc__btn_recover.html"
+  "include_html_file_name": "userprofile/includes/inc__btn_recover.html"
  }
 },
 {
@@ -174,7 +174,6 @@
   "author": 1,
   "is_active": true,
   "add_datetime": "2021-10-10T19:36:26Z",
-  "article_status": "MD",
   "article_status_new": 2
  }
 },
@@ -191,7 +190,6 @@
   "author": 1,
   "is_active": true,
   "add_datetime": "2021-10-10T19:37:51.106Z",
-  "article_status": "DR",
   "article_status_new": 1
  }
 },
@@ -208,7 +206,6 @@
   "author": 1,
   "is_active": true,
   "add_datetime": "2021-10-11T19:37:51.106Z",
-  "article_status": "AR",
   "article_status_new": 2
  }
 },
@@ -225,7 +222,6 @@
   "author": 1,
   "is_active": true,
   "add_datetime": "2021-10-11T19:37:51.106Z",
-  "article_status": "RC",
   "article_status_new": 3
  }
 },
@@ -242,7 +238,6 @@
   "author": 1,
   "is_active": true,
   "add_datetime": "2021-10-10T19:37:51.106Z",
-  "article_status": "PB",
   "article_status_new": 4
  }
 },
@@ -259,7 +254,6 @@
   "author": 1,
   "is_active": true,
   "add_datetime": "2021-10-10T19:36:26Z",
-  "article_status": "PB",
   "article_status_new": 4
  }
 },
@@ -276,7 +270,6 @@
   "author": 1,
   "is_active": true,
   "add_datetime": "2021-10-10T19:37:51.106Z",
-  "article_status": "PB",
   "article_status_new": 4
  }
 },
@@ -293,7 +286,6 @@
   "author": 1,
   "is_active": true,
   "add_datetime": "2021-10-10T19:36:26Z",
-  "article_status": "PB",
   "article_status_new": 4
  }
 },
@@ -310,7 +302,6 @@
   "author": 1,
   "is_active": true,
   "add_datetime": "2021-10-11T19:37:51.106Z",
-  "article_status": "PB",
   "article_status_new": 4
  }
 },
@@ -327,7 +318,6 @@
   "author": 1,
   "is_active": true,
   "add_datetime": "2021-10-11T19:37:51.106Z",
-  "article_status": "PB",
   "article_status_new": 4
  }
 },
@@ -344,7 +334,6 @@
   "author": 5,
   "is_active": true,
   "add_datetime": "2021-10-21T13:32:30.639Z",
-  "article_status": "PB",
   "article_status_new": 4
  }
 },
@@ -361,7 +350,6 @@
   "author": 5,
   "is_active": true,
   "add_datetime": "2021-10-21T13:33:59.323Z",
-  "article_status": "PB",
   "article_status_new": 4
  }
 },
@@ -378,7 +366,6 @@
   "author": 5,
   "is_active": true,
   "add_datetime": "2021-10-21T13:34:52.369Z",
-  "article_status": "PB",
   "article_status_new": 4
  }
 },
@@ -395,7 +382,6 @@
   "author": 2,
   "is_active": true,
   "add_datetime": "2021-10-21T13:36:47.171Z",
-  "article_status": "PB",
   "article_status_new": 4
  }
 },
@@ -412,7 +398,6 @@
   "author": 2,
   "is_active": true,
   "add_datetime": "2021-10-21T13:38:16.288Z",
-  "article_status": "PB",
   "article_status_new": 4
  }
 },
@@ -429,7 +414,6 @@
   "author": 3,
   "is_active": true,
   "add_datetime": "2021-10-21T13:40:31.853Z",
-  "article_status": "PB",
   "article_status_new": 4
  }
 },
@@ -446,7 +430,6 @@
   "author": 3,
   "is_active": true,
   "add_datetime": "2021-10-21T13:42:14.213Z",
-  "article_status": "PB",
   "article_status_new": 4
  }
 },
@@ -463,7 +446,6 @@
   "author": 3,
   "is_active": true,
   "add_datetime": "2021-10-21T13:43:46.492Z",
-  "article_status": "PB",
   "article_status_new": 4
  }
 },
@@ -480,7 +462,6 @@
   "author": 4,
   "is_active": true,
   "add_datetime": "2021-10-21T13:44:59.526Z",
-  "article_status": "PB",
   "article_status_new": 4
  }
 },
@@ -497,7 +478,6 @@
   "author": 4,
   "is_active": true,
   "add_datetime": "2021-10-21T13:45:52.980Z",
-  "article_status": "PB",
   "article_status_new": 4
  }
 },
@@ -514,7 +494,6 @@
   "author": 6,
   "is_active": true,
   "add_datetime": "2021-10-21T13:49:06.469Z",
-  "article_status": "PB",
   "article_status_new": 4
  }
 },
@@ -531,7 +510,6 @@
   "author": 6,
   "is_active": true,
   "add_datetime": "2021-10-21T13:51:01.779Z",
-  "article_status": "PB",
   "article_status_new": 4
  }
 },
@@ -548,7 +526,6 @@
   "author": 6,
   "is_active": true,
   "add_datetime": "2021-10-21T13:52:59.434Z",
-  "article_status": "PB",
   "article_status_new": 4
  }
 },
@@ -565,7 +542,6 @@
   "author": 6,
   "is_active": true,
   "add_datetime": "2021-10-21T13:54:03.581Z",
-  "article_status": "PB",
   "article_status_new": 4
  }
 },
@@ -582,7 +558,6 @@
   "author": 6,
   "is_active": true,
   "add_datetime": "2021-10-21T13:56:01.153Z",
-  "article_status": "PB",
   "article_status_new": 4
  }
 },
@@ -599,7 +574,6 @@
   "author": 5,
   "is_active": true,
   "add_datetime": "2021-10-21T13:57:40.263Z",
-  "article_status": "PB",
   "article_status_new": 4
  }
 },
@@ -616,7 +590,6 @@
   "author": 5,
   "is_active": true,
   "add_datetime": "2021-10-21T13:59:09.178Z",
-  "article_status": "PB",
   "article_status_new": 4
  }
 },
@@ -633,7 +606,6 @@
   "author": 5,
   "is_active": true,
   "add_datetime": "2021-10-21T14:01:09.103Z",
-  "article_status": "PB",
   "article_status_new": 4
  }
 },
@@ -650,7 +622,6 @@
   "author": 6,
   "is_active": true,
   "add_datetime": "2021-10-21T14:38:10.305Z",
-  "article_status": "PB",
   "article_status_new": 4
  }
 },
@@ -667,7 +638,6 @@
   "author": 6,
   "is_active": true,
   "add_datetime": "2021-10-21T14:39:38.974Z",
-  "article_status": "PB",
   "article_status_new": 4
  }
 },
@@ -684,7 +654,6 @@
   "author": 5,
   "is_active": true,
   "add_datetime": "2021-10-21T14:41:11.486Z",
-  "article_status": "PB",
   "article_status_new": 4
  }
 },
@@ -701,7 +670,6 @@
   "author": 5,
   "is_active": true,
   "add_datetime": "2021-10-21T14:42:06.005Z",
-  "article_status": "PB",
   "article_status_new": 4
  }
 },
@@ -718,7 +686,6 @@
   "author": 5,
   "is_active": true,
   "add_datetime": "2021-10-21T14:43:12.544Z",
-  "article_status": "PB",
   "article_status_new": 4
  }
 },
@@ -735,7 +702,6 @@
   "author": 5,
   "is_active": true,
   "add_datetime": "2021-10-21T14:46:33.124Z",
-  "article_status": "PB",
   "article_status_new": 4
  }
 },
@@ -752,7 +718,6 @@
   "author": 5,
   "is_active": true,
   "add_datetime": "2021-10-21T14:48:06.964Z",
-  "article_status": "PB",
   "article_status_new": 4
  }
 },
@@ -769,7 +734,6 @@
   "author": 5,
   "is_active": true,
   "add_datetime": "2021-10-21T14:49:03.866Z",
-  "article_status": "PB",
   "article_status_new": 4
  }
 },
