+++ mainapp/context_processors.py
@@ -1,4 +1,4 @@
-from mainapp.models import Hub, Article
+from mainapp.models import Hub, Article, ArticleStatus
 from mainapp.search_filter import ArticleFilter
 
 
@@ -7,6 +7,6 @@ def category(request):
 
 
 def search_filter(request):
-    article = Article.objects.filter(article_status='PB')
+    article = Article.objects.filter(article_status_new=ArticleStatus.objects.get(name='Опубликована'))
     search_filter = ArticleFilter(request.GET, queryset=article)
     return {'search_filter': search_filter}
