+++ mainapp/models.py
@@ -21,16 +21,15 @@ class Hub(models.Model):
         return f'{self.name}'
 
 
-class ButtonsInProfile(models.Model):
-    name = models.CharField(verbose_name='название кнопки', max_length=20)
-    include_html_file_name = models.CharField(verbose_name='html файл', max_length=100)
-
-
 class ArticleStatus(models.Model):
     short_name = models.CharField(verbose_name='Краткое название', max_length=2)
     name = models.CharField(verbose_name='Название статуса', max_length=50)
     name_plural = models.CharField(verbose_name='Название раздела в ЛК (мн.число)', max_length=50)
-    buttons = models.ManyToManyField(ButtonsInProfile, verbose_name='Кнопки для этого статуса в ЛК')
+    buttons = models.ManyToManyField(
+        ButtonsInProfile,
+        verbose_name='Кнопки для этого статуса в ЛК',
+        related_name='кнопки'
+    )
 
 
 class Article(models.Model):
@@ -80,12 +79,6 @@ class Article(models.Model):
         verbose_name='Время добавления',
         auto_now_add=True
     )
-    article_status = models.CharField(
-        verbose_name='Статус публикации',
-        max_length=2,
-        choices=ARTICLE_STATUS_CHOICES,
-        default=ARTICLE_DEFAULT_STATUS
-    )
     article_status_new = models.ForeignKey(
         ArticleStatus,
         verbose_name='Статус публикации (новый)',
@@ -93,12 +86,7 @@ class Article(models.Model):
         on_delete=models.SET_NULL,
     )
     views = models.IntegerField(default=0, verbose_name='просмотры')
-    article_status_new = models.ForeignKey(
-        ArticleStatus,
-        verbose_name='Статус публикации (новый)',
-        null=True,
-        on_delete=models.SET_NULL,
-    )
+    rating = models.IntegerField(default=0, verbose_name='рейтинг')
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
