+++ mainapp/templates/mainapp/articles.html
@@ -10,7 +10,7 @@
                     <div class="related-articles">
                         {% for item in object_list %}
                             <div class="block-article">
-                                <p><a href="{% url 'auth:profile' %}">
+                                <p><a href="{% url 'profile:user' item.author.pk %}">
                             <span class="font-weight-bold"><img class="mr-2"
                                                                 src="{{ item.author.avatar|media_folder_users }}"
                                                                 alt="profile_image" height="24">
