+++ mainapp/templates/mainapp/includes/inc__activity.html
@@ -52,7 +52,7 @@
                 <div class="block-comment">
                     <div class="text" name="{{item.id}}">
                         <p style="color: maroon">
-                            <a href="{% url 'auth:profile' %}">
+                            <a href="{% url 'profile:main' %}">
                                 <span class="font-weight-bold">
                                     {{ item.author }}
                                 </span>
