+++ mainapp/templates/mainapp/includes/inc__main_menu.html
@@ -91,7 +91,7 @@
                     {% endif %}
                 </a>
                     <div class="menu-icon-box">
-                        <a id="menu-icon" class="user-panel-button" title="Личный кабинет" href="{% url 'auth:profile' %}">
+                        <a id="menu-icon" class="user-panel-button" title="Личный кабинет" href="{% url 'profile:main' %}">
                         <svg width="25" height="25" viewBox="0 0 16 16" class="bi bi-person-fill" fill="currentColor"
                              xmlns="http://www.w3.org/2000/svg">
                             <path fill-rule="evenodd"
